requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
anthropic==0.43.0
networkx==3.4.2
structlog==24.4.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
